refacto: refactorizacion de codigo masivo
