+++ src/bot/parser/Parser.interfaces.ts
@@ -1,22 +1,27 @@
-import { info_calificaciones, info_contacto, info_estudiante, info_extension, info_extra, info_habilitacion_actual, info_horario_clase, info_horario_docente, info_inscripciones_asistencia, info_libros_prestamo, info_libros_reservas, info_materia_pendiente, info_resultado_evaluacion_final, info_resultado_parcial, info_tiempo_rendimiento, info_ultimos_pagos } from "../../types/ConsultorInfoTipos/ConsultorEstudianteTipos";
 
-export interface TableContent {
-    headers: string[];
-    rows: string[][];
-}
-export interface TableContentObjects {
-    [key: string]:TableContent;
-}
+import { info_estudiante, 
+    info_contacto, 
+    info_tiempo_rendimiento,
+    info_inscripciones_asistencia, 
+    info_ultimos_pagos, 
+    info_resultado_parcial, 
+    info_habilitacion_actual, 
+    info_resultado_evaluacion_final, 
+    info_calificaciones, 
+    info_materia_pendiente, 
+    info_extension, 
+    info_horario_clase, 
+    info_horario_docente, 
+    info_libros_prestamo, 
+    info_libros_reservas
+ } from "../../types/ConsultorEstudiante.types";
 
-export interface ITableContentPreProcessor {
-    get_TableConsultorObjects(): TableContentObjects;
-}
-export interface IParserInfoEstudiante {
+export interface IBasicEstudianteInfoParser {
     get_info_estudiante():info_estudiante;
     get_info_contacto():info_contacto;
     get_info_tiempo_rendimiento():info_tiempo_rendimiento;
 }
-export interface IConsultorParserInterface extends IParserInfoEstudiante {
+export interface IEstudianteInfoTableParser {
     get_info_inscipciones_asistencia(): info_inscripciones_asistencia[];
     get_info_ultimos_pagos():info_ultimos_pagos[];
     get_info_resultados_parciales(): info_resultado_parcial[];
