+++ src/bot/processors/processors.interfaces.ts
@@ -1,0 +1,6 @@
+import { TableContentObjects } from "./processors.types";
+
+
+export interface ITableContentProcessor{
+    process():Promise<TableContentObjects>;
+}
