+++ src/bot/processors/processors.types.ts
@@ -1,0 +1,8 @@
+
+export interface TableContent {
+    headers: string[];
+    rows: string[][];
+}
+export interface TableContentObjects {
+    [key: string]:TableContent;
+}
