+++ src/types/ConsultorEstudianteVistas.types.ts
@@ -1,6 +1,4 @@
-import {info_estudiante} from './ConsultorEstudianteTipos';
-
-
+import {info_estudiante} from './ConsultorEstudiante.types';
 
 export interface vista_info_consultor {
     info_estudiante_cabecera:info_estudiante
