--- src/types/Credential.ts
@@ -1,4 +1,0 @@
-export interface LoginCredentials{
-    password : string , 
-    username: string
-}
